Update show.php with Category IF-ELSEIF
Added IF-ELSEIF statements based on Category value (and the metadata to show for each case). Also added Type metadata display for the "formation" or cultural association of an instrument (same line as the Spatial Coverage metadata for "nation"). Also changed the delimiter for Coverage metadata; for now left this as one metadata field, though Roger may want it split into two input fields.